dev > main: refactorisation
